feat: Set password of Realm-Property-User via Callback
The password is provided via a callback, as the realm.properties contains the password in plain text. However, the User:setPassword() expects a hashed password. hashing the password is very time-consuming. The callback ensures that the password is only hashed if it is necessary. It is only necessary for an authentication process. It is not necessary when creating a user object via a JWT. This optimization is precisely for this case.